requirements.txt
Flask==3.1.0
Pillow==11.0.0
opencv-python==4.10.0.84
rembg>=2.0.0
qrcode>=7.4.2
numpy==2.0.0